% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bayes_factor_kodikara2020.R
\name{compute_bayes_factor_kodikara2020}
\alias{compute_bayes_factor_kodikara2020}
\title{Compute Bayes factor following a non-homogeneous Poisson process.}
\usage{
compute_bayes_factor_kodikara2020(
  sightings,
  start_year,
  end_year,
  dprior_m,
  dprior_te
)
}
\arguments{
\item{sightings}{sequence of years with observations in a vector}

\item{start_year}{starting study period}

\item{end_year}{ending study period}

\item{dprior_m}{function to estimate the prior observation rate m}

\item{dprior_te}{function to estimate the prior extinction time}
}
\value{
Bayes factor comparing persistence to extinction.
}
\description{
Compute Bayes factor following a non-homogeneous Poisson process.
}
\examples{
compute_bayes_factor_kodikara2020(sightings = c(1901,1902,1903,1905,1908,1910), start_year = 1900, end_year = 1920, dprior_m = solowdprior_m, dprior_te = solowdprior_te)
}
