% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_bayes_factor_kodikara2020_fixed_alpha.R
\name{compute_bayes_factor_kodikara2020_fixed_alpha}
\alias{compute_bayes_factor_kodikara2020_fixed_alpha}
\title{Compute bayes factor kodikara2020 fixed alpha}
\usage{
compute_bayes_factor_kodikara2020_fixed_alpha(
  sightings,
  start_year,
  end_year,
  dprior_m,
  dprior_te,
  alpha
)
}
\arguments{
\item{sightings}{Numeric vector of sighting years.}

\item{start_year}{First year of the study period.}

\item{end_year}{Last year of the study period.}

\item{dprior_m}{Function for the prior distribution of observation rate.}

\item{dprior_te}{Function for the prior distribution of extinction time.}

\item{alpha}{Fixed shape parameter for the non-homogeneous process.}
}
\value{
Bayes factor comparing persistence to extinction.
}
\description{
Compute bayes factor kodikara2020 fixed alpha
}
