% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bayes_factor_solow1993.R
\name{compute_bayes_factor_solow1993}
\alias{compute_bayes_factor_solow1993}
\title{Computes the Bayes factor from Solow 1993}
\usage{
compute_bayes_factor_solow1993(
  sightings,
  start_year,
  end_year,
  dprior_m,
  dprior_te
)
}
\arguments{
\item{sightings}{input vector with years}

\item{start_year}{numeric value that equals to 1 year before the first sighting}

\item{end_year}{numeric value of the ending  study period}

\item{dprior_m}{function of the prior observation rate m}

\item{dprior_te}{function of the prior extinction time}
}
\value{
The Bayes factor between the species still extant / species going extinct
}
\description{
Computes the Bayes factor from Solow 1993
}
\examples{
compute_bayes_factor_solow1993(sightings = c(1901,1902,1903,1905,1908,1910), start_year = 1900, end_year = 1920, dprior_m = function(m) 1 / m, dprior_te = function(te) 1)

}
