% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_likelihood_extinction_at_te_kodikara.R
\name{compute_likelihood_extinction_at_te_kodikara}
\alias{compute_likelihood_extinction_at_te_kodikara}
\alias{compute_likelyhood_extinction_at_te_kodikara}
\title{Compute extinction likelihood at an extinction time}
\usage{
compute_likelihood_extinction_at_te_kodikara(t, te, dprior_m)

compute_likelyhood_extinction_at_te_kodikara(t, te, dprior_m)
}
\arguments{
\item{t}{number of sightings}

\item{te}{extinction time}

\item{dprior_m}{function of the prior observation rate m}
}
\value{
extinction likelihood at a specific extinction time
}
\description{
Compute extinction likelihood at an extinction time
}
