% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_likelihood_extinction_at_te_kodikara_fixed_alpha.R
\name{compute_likelihood_extinction_at_te_kodikara_fixed_alpha}
\alias{compute_likelihood_extinction_at_te_kodikara_fixed_alpha}
\alias{compute_likelyhood_extinction_at_te_kodikara_fixed_alpha}
\title{Likelihood of Extinction at a Fixed Alpha Parameter (Kodikara Method)}
\usage{
compute_likelihood_extinction_at_te_kodikara_fixed_alpha(
  t,
  te,
  dprior_m,
  alpha
)

compute_likelyhood_extinction_at_te_kodikara_fixed_alpha(
  t,
  te,
  dprior_m,
  alpha
)
}
\arguments{
\item{t}{Numeric vector of scaled sighting times.}

\item{te}{Scaled extinction time.}

\item{dprior_m}{Function for the prior distribution of observation rate.}

\item{alpha}{Fixed shape parameter for the non-homogeneous process.}
}
\value{
Likelihood of extinction at \code{te}.
}
\description{
Likelihood of Extinction at a Fixed Alpha Parameter (Kodikara Method)
}
