% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_likelihood_at_te.R
\name{compute_likelyhood_extinction_at_te}
\alias{compute_likelyhood_extinction_at_te}
\title{Compute likelihood of extinction at te}
\usage{
compute_likelyhood_extinction_at_te(t, te, dprior_m)
}
\arguments{
\item{t}{number of sightings}

\item{te}{extinction time}

\item{dprior_m}{function of the prior observation rate m}
}
\value{
likelihood of extinction at extinction time (te)
}
\description{
Compute likelihood of extinction at te
}
\examples{
compute_likelyhood_extinction_at_te(t = (c(1901,1902,1903,1905,1908,1910) - 1900) / (1920 - 1900), 1, dprior_m = function(m) 1 / m)
}
