% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_c2022_extinction.R
\name{compute_posterior_c2022_extinction}
\alias{compute_posterior_c2022_extinction}
\title{Compute  Posterior Extinction with informative priors}
\usage{
compute_posterior_c2022_extinction(
  sightings,
  start_year,
  stop_year,
  prior_te = NULL,
  prior_m = NULL
)
}
\arguments{
\item{sightings}{list of observations}

\item{start_year}{first year of study period}

\item{stop_year}{last year of study period}

\item{prior_te}{flexible function - add prior extinction rate distribution}

\item{prior_m}{flexible function - add prior observation rate distribution}
}
\value{
Numeric vector of posterior extant probabilities by year.
}
\description{
Compute  Posterior Extinction with informative priors
}
