% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_from_bayes_solow1993.R
\name{compute_posterior_from_bayes_solow1993}
\alias{compute_posterior_from_bayes_solow1993}
\title{Compute  Posterior  from bayes solow1993}
\usage{
compute_posterior_from_bayes_solow1993(bayes_factor, prior)
}
\arguments{
\item{bayes_factor}{Automatically calculated}

\item{prior}{usually is 0.5}
}
\value{
Posterior  from bayes solow1993
}
\description{
Compute  Posterior  from bayes solow1993
}
\examples{

}
