% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_non_homogeneous.R
\name{compute_posterior_non_homogeneous}
\alias{compute_posterior_non_homogeneous}
\title{Compute Posterior using Non homogeneous observation rate}
\usage{
compute_posterior_non_homogeneous(
  sightings,
  start_year,
  end_year,
  dprior_m,
  dprior_te,
  prior = 0.5
)
}
\arguments{
\item{sightings}{sequence of years with observations in a vector}

\item{start_year}{starting study period}

\item{end_year}{ending study period}

\item{dprior_m}{function to estimate the prior observation rate m}

\item{dprior_te}{function to estimate the prior extinction time}

\item{prior}{extinction time probability, usually fix to 0.5 but can be between 0 and 1}
}
\value{
Posterior distribution of the extant probability of a species during the study period
}
\description{
Compute Posterior using Non homogeneous observation rate
}
\examples{
compute_posterior_non_homogeneous(sightings = c(1901,1902,1903,1905,1908,1910), start_year = 1900, end_year = 1920, dprior_m = solowdprior_m,  dprior_te = solowdprior_te, prior = 0.5)
}
