% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_posterior_non_homogeneous_fixed_alpha.R
\name{compute_posterior_non_homogeneous_fixed_alpha}
\alias{compute_posterior_non_homogeneous_fixed_alpha}
\title{Compute Posterior non homogeneous fixed alpha}
\usage{
compute_posterior_non_homogeneous_fixed_alpha(
  sightings,
  start_year,
  end_year,
  dprior_m,
  dprior_te,
  prior = 0.5,
  alpha
)
}
\arguments{
\item{sightings}{Numeric vector of sighting years.}

\item{start_year}{First year of the study period.}

\item{end_year}{Last year of the study period.}

\item{dprior_m}{Function for the prior distribution of observation rate.}

\item{dprior_te}{Function for the prior distribution of extinction time.}

\item{prior}{Prior probability that the species is extant.}

\item{alpha}{Fixed shape parameter for the non-homogeneous process.}
}
\value{
Posterior probability that the species is extant.
}
\description{
Compute Posterior non homogeneous fixed alpha
}
