% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_m.R
\name{dprior_m}
\alias{dprior_m}
\title{Solow 1993 prior function for m}
\usage{
dprior_m(m)
}
\arguments{
\item{m}{}
}
\value{
prior function for m
}
\description{
Solow 1993 prior function for m
}
\examples{
NOT APPLY
}
