% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_te.R
\name{dprior_te}
\alias{dprior_te}
\title{Solow 1993 Prior function for te}
\usage{
dprior_te(te)
}
\arguments{
\item{te}{}
}
\value{
Prior function for te
}
\description{
Solow 1993 Prior function for te
}
\examples{
NOT APPLY

}
