% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dependence.R
\name{final_year_probability_grid}
\alias{final_year_probability_grid}
\title{Extract the final posterior probability from each spatial cell}
\usage{
final_year_probability_grid(spatial_posterior)
}
\arguments{
\item{spatial_posterior}{Matrix/list of posterior probability vectors.}
}
\value{
Numeric matrix with the final posterior probability for each cell.
}
\description{
Extract the final posterior probability from each spatial cell
}
