% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dependence.R
\name{make_grid_adjacency}
\alias{make_grid_adjacency}
\title{Build an adjacency list for a regular spatial grid}
\usage{
make_grid_adjacency(nrow, ncol, diagonal = FALSE)
}
\arguments{
\item{nrow}{Number of grid rows.}

\item{ncol}{Number of grid columns.}

\item{diagonal}{Logical. If \code{TRUE}, diagonal neighbours are included.}
}
\value{
A list of integer vectors. Each element contains the linear indices
of the neighbours of one grid cell.
}
\description{
Build an adjacency list for a regular spatial grid
}
\examples{
make_grid_adjacency(2, 2)
}
