% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extant_probability_by_yearsnonhomogeneous.R
\name{Posterior_probability_extinction_non_homogeneos_varying_end_year}
\alias{Posterior_probability_extinction_non_homogeneos_varying_end_year}
\title{Extant Posterior Probability sequence by years folowing non homogeneous}
\usage{
Posterior_probability_extinction_non_homogeneos_varying_end_year(
  sightings,
  start_year,
  stop_year
)
}
\arguments{
\item{sightings}{Vector with the years with  observations}

\item{start_year}{Year at which we want to estimate the extant posterior probability}

\item{stop_year}{Year at which we stop to estimate the extant  posterior probability}
}
\value{
Vector with the extant probabilities
}
\description{
Extant Posterior Probability sequence by years folowing non homogeneous
}
