% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extant_probability_by_years_non_homogeneous.R
\name{posterior_probability_extinction_non_homogeneous_varying_end_year}
\alias{posterior_probability_extinction_non_homogeneous_varying_end_year}
\alias{posterior_probability_extinction_non_homogeneos_varying_end_year}
\title{Extant posterior probability sequence by year following a non-homogeneous process}
\usage{
posterior_probability_extinction_non_homogeneous_varying_end_year(
  sightings,
  start_year,
  stop_year
)

posterior_probability_extinction_non_homogeneos_varying_end_year(
  sightings,
  start_year,
  stop_year
)
}
\arguments{
\item{sightings}{Vector with the years with  observations}

\item{start_year}{Year at which we want to estimate the extant posterior probability}

\item{stop_year}{Year at which we stop to estimate the extant  posterior probability}
}
\value{
Vector with the extant probabilities
}
\description{
Extant posterior probability sequence by year following a non-homogeneous process
}
