% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first_n_observations.R
\name{select_first_n_observations}
\alias{select_first_n_observations}
\title{Select the first n observations to retain}
\usage{
select_first_n_observations(sightings, n)
}
\arguments{
\item{sightings}{Numeric vector of sighting years.}

\item{n}{Number of observations to keep.}
}
\value{
The first \code{n} observations from \code{sightings}.
}
\description{
Select the first n observations to retain
}
