% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solow2003_function.R
\name{solow2003_function}
\alias{solow2003_function}
\title{solow2003_function}
\usage{
solow2003_function(sightings, alpha)
}
\arguments{
\item{sightings}{Numeric vector of sighting years.}

\item{alpha}{Significance level used to compute the confidence interval.}
}
\value{
solow2003_function - estimated
}
\description{
solow2003_function
}
