% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_te.R
\name{solowdprior_te}
\alias{solowdprior_te}
\title{Prior extinction time from solow 1993}
\usage{
solowdprior_te(te)
}
\arguments{
\item{te}{extinction time}
}
\value{
function prior extinction time
}
\description{
Prior extinction time from solow 1993
}
