% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaColExt-package.R
\docType{package}
\name{SpaColExt-package}
\alias{SpaColExt}
\alias{SpaColExt-package}
\title{SpaColExt: Infer spatial colonization and extinction dates}
\description{
SpaColExt implements Bayesian estimators for extinction inference from
sighting records, including Solow-style estimators and spatial wrappers for
gridded sighting data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Clara-Casabona/SpaColExt}
  \item Report bugs at \url{https://github.com/Clara-Casabona/SpaColExt/issues}
}

}
\author{
\strong{Maintainer}: Clara Casabona i Amat \email{Clara.Casabona.I.Amat@USherbrooke.ca}

Authors:
\itemize{
  \item Clara Casabona i Amat \email{Clara.Casabona.I.Amat@USherbrooke.ca}
  \item Simon Bouchard
}

}
\keyword{internal}
