% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/spatial_posterior_probability_extinction_varying_end_year.R
\name{spatial_posterior_probability_extinction_varying_end_year}
\alias{spatial_posterior_probability_extinction_varying_end_year}
\title{Spatial Extant Posterior Probability  by years}
\usage{
spatial_posterior_probability_extinction_varying_end_year(
  sightings,
  start_year,
  stop_year
)
}
\arguments{
\item{sightings}{Matrix with vectors that contain the years with  observations}

\item{start_year}{Year at which we want to estimate the extant posterior probability}

\item{stop_year}{Year at which we stop to estimate the extant  posterior probability}
}
\value{
Matrix/list of posterior extant probability vectors by site.
}
\description{
Spatial Extant Posterior Probability  by years
}
\examples{
data = list(c(1880, 1883, 1895, 1897, 1899), NA, NA, c(1882, 1884, 1896, 1898))
dim(data) <- c(2, 2)
spatial_posterior_probability_extinction_varying_end_year(data, start_year=1880, stop_year=1920)


}
