% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_and_reverse.R
\name{transform_and_reverse}
\alias{transform_and_reverse}
\title{Transform and reverse sighting dates for colonisation models}
\usage{
transform_and_reverse(sightings, t_start, t_end)
}
\arguments{
\item{sightings}{Numeric vector of sighting years.}

\item{t_start}{First year of the study period.}

\item{t_end}{Last year of the study period.}
}
\value{
inverted sightings
}
\description{
Transform and reverse sighting dates for colonisation models
}
\examples{
transform_and_reverse(c(1901, 1902, 1903, 1905, 1908, 1910), 1900, 1920)
}
