% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_priors_effects.R
\name{visualize_priors_effects}
\alias{visualize_priors_effects}
\alias{vizualize_priors_effects}
\title{Visualize the effect of prior distributions}
\usage{
visualize_priors_effects(t_start, t_stop, sightings, prior_te, prior_m)

vizualize_priors_effects(t_start, t_stop, sightings, prior_te, prior_m)
}
\arguments{
\item{t_start}{First year of the study period.}

\item{t_stop}{Last year of the study period.}

\item{sightings}{Numeric vector of sighting years.}

\item{prior_te}{Function for the prior distribution of extinction time.}

\item{prior_m}{Function for the prior distribution of observation rate.}
}
\value{
A \code{ggplot} object comparing posterior curves under different priors.
}
\description{
Visualize the effect of prior distributions
}
\examples{
sightings <- c(1901, 1902, 1903, 1905, 1908, 1910)
prior_te <- function(te) 1
prior_m <- function(m) 1 / pmax(m, .Machine$double.eps)
visualize_priors_effects(1900, 1920, sightings, prior_te, prior_m)
}
