% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_priors_effects.R
\name{vizualize_priors_effects}
\alias{vizualize_priors_effects}
\title{Vizualize Priors Effects}
\usage{
vizualize_priors_effects(t_start, t_stop, sightings, prior_te, prior_m)
}
\arguments{
\item{t_start}{Starting study period}

\item{t_stop}{End of the study period}

\item{sightings}{Vector with the years of observation}

\item{prior_te}{Function with the prior extinction distribution}

\item{prior_m}{Function with the prior observation rate distribution}
}
\value{

}
\description{
Vizualize Priors Effects
}
