% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_c2022_extinction.R
\name{compute_posterior_c2022_extinction}
\alias{compute_posterior_c2022_extinction}
\title{Compute Posterior extant probability with new TE prior}
\usage{
compute_posterior_c2022_extinction(
  sightings,
  start_year,
  stop_year,
  extinction_posteriorLevel
)
}
\arguments{
\item{sightings}{matrix array with years by cell}

\item{start_year}{Start of the study period}

\item{stop_year}{End of the study period}

\item{extinction_posteriorLevel}{Posterior extant probability of a higher spatial level analysis}
}
\value{
the posterior extrant probability of an species by location
}
\description{
Compute Posterior extant probability with new TE prior
}
\examples{

}
