% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_from_bayes_solow1993.R
\name{compute_posterior_from_bayes_solow1993}
\alias{compute_posterior_from_bayes_solow1993}
\title{Compute  Posterior  from bayes solow1993}
\usage{
compute_posterior_from_bayes_solow1993(bayes_factor, prior)
}
\arguments{
\item{bayes_factor}{Automatically calculated}

\item{prior}{usually is 0.5}
}
\value{
Posterior  from bayes solow1993
}
\description{
Compute  Posterior  from bayes solow1993
}
\examples{
compute_posterior_from_bayes_solow1993(compute_bayes_factor_solow1993(sightings = c(1901,1902,1903,1905,1908,1910) ,start_year = 1900, end_year = 1920,dprior_m = function(m) 1 / m, dprior_te = function(te) 1),0.5)

}
