% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_posterior_distribution.R
\name{Plot_posterior_distribution}
\alias{Plot_posterior_distribution}
\title{Plot temporal posterior distribution}
\usage{
Plot_posterior_distribution(extinction_posterior, start_year, stop_year)
}
\arguments{
\item{extinction_posterior}{Output from the posterior probability extinction functions (vector or list of vectors)}

\item{start_year}{Year at which we want to estimate the extant posterior probability}

\item{stop_year}{Year at which we stop to estimate the extant  posterior probability}

\item{year}{Variable generated automaticaly in the function to scale the plot}

\item{value}{Corresponds to the extinction posterior value}
}
\value{
Plot with posterior extant probability
}
\description{
Plot temporal posterior distribution
}
\examples{
data = list(c(1880, 1883, 1895, 1897, 1899), NA, NA, c(1882, 1884, 1896, 1898))
dim(data) <- c(2, 2)
Plot_posterior_distribution(Spatial_posterior_probability_extinction_varying_end_year(data, start_year=1880, stop_year=1920), start_year=1880, stop_year=1920)
}
