% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Posterior_probability_extinction_varing_end_year.R
\name{Posterior_probability_extinction_varying_end_year}
\alias{Posterior_probability_extinction_varying_end_year}
\title{Extant Posterior Probability sequence by years}
\usage{
Posterior_probability_extinction_varying_end_year(
  sightings,
  start_year,
  stop_year
)
}
\arguments{
\item{sightings}{Vector with the years with  observations}

\item{start_year}{Year at which we want to estimate the extant posterior probability}

\item{stop_year}{Year at which we stop to estimate the extant  posterior probability}
}
\value{
Vector with the extant probabilities
}
\description{
Extant Posterior Probability sequence by years
}
\examples{
Posterior_probability_extinction_varying_end_year(sightings = c(1898, 1899, 1900, 1901, 1902, 1904, 1905, 1906, 1907, 1908, 1909, 1910, 1913),1898, 1930)
}
